+++ packages/use-analytics/src/analytics/useAnalytics.tsx
@@ -5,8 +5,12 @@ import type { ReactNode } from 'react'
 import { useDeepCompareEffectNoCheck } from 'use-deep-compare-effect'
 import { destSDKBaseURL, pluginsSDKBaseURL } from '../constants'
 import type { CategoryKind } from '../types'
-import { defaultConsentOptions, defaultLoadOptions } from './constants'
-import { userMigrationsTraits } from './segments/userMigrationsTraits'
+import {
+  defaultConsentOptions,
+  defaultLoadOptions,
+  defaultTimeout,
+} from './constants'
+import { normalizeIdsMigration } from './normalizeIdsMigration'
 
 type Analytics = RudderAnalytics
 export type { Analytics }
@@ -50,7 +54,7 @@ export type AnalyticsProviderProps<T> = {
   loadOptions?: LoadOptions
 
   /**
-   *  This option force provider to render children only when isAnalytics is ready
+   *  This option force provider to render children only when isAnalytics is ready, you can also set a timeout to prevent blocking indefinitely and use isAnalyticsReady to show a loading screen.
    */
   shouldRenderOnlyWhenReady?: boolean
   /**
@@ -92,11 +96,18 @@ export function AnalyticsProvider<T extends Events>({
   // This effect will unlock the case where we have a failure with the load of the analytics.load as rudderstack doesn't provider any solution for this case.
   useEffect(() => {
     let timer: ReturnType<typeof setTimeout> | undefined
-    if (!isAnalyticsReady && !internalAnalytics && timeout) {
-      if (shouldRenderOnlyWhenReady) {
-        timer = setTimeout(() => setIsAnalyticsReady(true), timeout)
-        onError?.(new Error('Analytics Setup Timeout'))
-      }
+    if (
+      !isAnalyticsReady &&
+      (Number.isFinite(timeout) || shouldRenderOnlyWhenReady)
+    ) {
+      timer = setTimeout(() => {
+        setIsAnalyticsReady(true)
+        onError?.(new Error('Timeout'))
+      }, timeout ?? defaultTimeout)
+    }
+
+    if (isAnalyticsReady) {
+      clearTimeout(timer)
     }
 
     return () => {
@@ -104,7 +115,6 @@ export function AnalyticsProvider<T extends Events>({
     }
   }, [
     isAnalyticsReady,
-    internalAnalytics,
     setIsAnalyticsReady,
     shouldRenderOnlyWhenReady,
     timeout,
@@ -129,8 +139,6 @@ export function AnalyticsProvider<T extends Events>({
         destSDKBaseURL: destSDKBaseURL(settings.cdnURL),
         pluginsSDKBaseURL: pluginsSDKBaseURL(settings.cdnURL),
         onLoaded: (rudderAnalytics: Analytics) => {
-          userMigrationsTraits(rudderAnalytics)
-
           rudderAnalytics.consent({
             ...defaultConsentOptions,
             consentManagement: {
@@ -139,16 +147,22 @@ export function AnalyticsProvider<T extends Events>({
               deniedConsentIds: deniedConsents,
             },
           })
+          normalizeIdsMigration(analytics)
 
           onLoaded(rudderAnalytics)
-
-          setIsAnalyticsReady(true)
+          setAnalytics(analytics)
         },
         ...loadOptions,
       })
 
       analytics.ready(() => {
-        setAnalytics(analytics)
+        /**
+         * this will wait for client destination to be ready, but analytics is ready.
+         * we can listen for RSA_Ready event to know when the analytics is ready if we don't want to wait for the client destination to be ready.
+         * document.addEventListener('RSA_Ready', function(e) {
+         *  console.log('RSA_Ready', e.detail.analyticsInstance);
+         * });
+         */
         setIsAnalyticsReady(true)
       })
     }