+++ packages/use-i18n/src/formatUnit.ts
@@ -159,7 +159,7 @@ const format =
           round: maximumFractionDigits,
         })
 
-        computedValue = Number.parseFloat(value.value)
+        computedValue = Number.parseFloat(value.value.toString())
       } else {
         const value = filesize(amount, {
           base,
@@ -168,7 +168,7 @@ const format =
         })
 
         computedExponent = exponents[value.exponent]
-        computedValue = Number.parseFloat(value.value)
+        computedValue = Number.parseFloat(value.value.toString())
       }
     }
 
@@ -194,9 +194,9 @@ const format =
 
 type SimpleUnits = `${ExponentName}${Unit}${'-humanized' | ''}`
 type ComplexUnits = `${Unit}${'s' | ''}${'-humanized' | ''}`
-type PerSecondUnit = `${ExponentName | ''}bit${'s' | ''}${'-per-second' | ''}${
-  | '-humanized'
-  | ''}`
+
+type PerSecondUnit =
+  `${ExponentName | ''}bit${'s' | ''}${'-per-second' | ''}${'-humanized' | ''}`
 type SupportedUnits = SimpleUnits | ComplexUnits | PerSecondUnit
 
 export const supportedUnits: Partial<